reshuffled test cases to make everything more reusable #9
