+++ test/test_clustering.py
@@ -1,16 +1,19 @@
-import unittest
-import tempfile
 import csv
+import unittest
 import os
+import tempfile
 from work_set_clustering.clustering import clusterFromScratch as initialClustering
 from work_set_clustering.clustering import updateClusters as updateClusters
+from test.test_cases import InitialClusteringSize, InitialElementsTogether, UpdateClusteringSize, UpdateClusteringElementsTogether
 
 # Don't show the traceback of an AssertionError, because the AssertionError already says what the issue is!
 __unittest = True
 
 # ---------------------------------------------------------------------------
 def readOutput(filename):
- with open(filename, 'r') as fileIn:
+  """This helper function reads the output of a clusterin run and creates a data structure with cluster assignments."""
+
+  with open(filename, 'r') as fileIn:
     csvReader = csv.DictReader(fileIn, delimiter=',')
 
     data = {
@@ -33,70 +36,8 @@ def readOutput(filename):
 
     return data
 
-
 # -----------------------------------------------------------------------------
-class TestClustering():
-  """A class with integration test cases for different implementations."""
-
-  # ---------------------------------------------------------------------------
-  def testCorrectNumberOfClusters(self):
-    """With given cluster input, two clusters should be found"""
-    numberFoundClusters = len(self.getInitialClusterData()['clusterIdentifiers'])
-    numberExpectedClusters = 2
-    self.assertEqual(numberFoundClusters, numberExpectedClusters, msg=f'Found {numberFoundClusters} clusters instead of {numberExpectedClusters}')
-
-  # ---------------------------------------------------------------------------
-  def testElement1And2Together(self):
-    """Element e1 and e2 should be clustered together"""
-    clusterE1 = self.getInitialClusterData()['elementToCluster']['e1']
-    clusterE2 = self.getInitialClusterData()['elementToCluster']['e2']
-    self.assertEqual(clusterE1, clusterE2, msg=f'Different clusters for e1 and e2 ({clusterE1} != {clusterE2})')
-
-  # ---------------------------------------------------------------------------
-  def testElement3And4Together(self):
-    """Element e3 and e4 should be clustered together"""
-    clusterE3 = self.getInitialClusterData()['elementToCluster']['e3']
-    clusterE4 = self.getInitialClusterData()['elementToCluster']['e4']
-    self.assertEqual(clusterE3, clusterE4, msg=f'Different clusters for e3 and e4 ({clusterE3} != {clusterE4})')
-
-  # ---------------------------------------------------------------------------
-  def testElement1And5Together(self):
-    """Element e5 should be clustered together with the initial e1 and e2"""
-    clusterInitial = self.getUpdatedClusterData()['elementToCluster']['e1']
-    clusterNew = self.getUpdatedClusterData()['elementToCluster']['e5']
-    self.assertEqual(clusterInitial, clusterNew, msg=f'Different clusters for initial e1 and updated e5 ({clusterInitial} != {clusterNew})')
-
-  # ---------------------------------------------------------------------------
-  def testElement1SameClusterAfterUpdate(self):
-    """Element e1 should have the same clusterID after update when e5 was added"""
-    clusterInitial = self.getUpdatedClusterData()['elementToCluster']['e1']
-    clusterNew = "c1"
-    self.assertEqual(clusterInitial, clusterNew, msg=f'Initial cluster of e1 has changed after update ({clusterInitial} != {clusterNew})')
-
-  # ---------------------------------------------------------------------------
-  def testElement5SameClusterAfterUpdate(self):
-    """Element e5 should be in the same clusterID as the initial e1"""
-    clusterID = self.getUpdatedClusterData()['elementToCluster']['e5']
-    clusterExpected = "c1"
-    self.assertEqual(clusterID, clusterExpected, msg=f'Element e5 was not added to the existing clusterID of e1 ({clusterID} != {clusterExpected})')
-
-  # ---------------------------------------------------------------------------
-  def testElement2And5Together(self):
-    """Element e5 should be clustered together with the initial e1 and e2"""
-    clusterInitial = self.getUpdatedClusterData()['elementToCluster']['e2']
-    clusterNew = self.getUpdatedClusterData()['elementToCluster']['e5']
-    self.assertEqual(clusterInitial, clusterNew, msg=f'Different clusters for initial e2 and updated e5 ({clusterInitial} != {clusterNew})')
-
-  # ---------------------------------------------------------------------------
-  def testElement7InNewCluster(self):
-    """Element e7 should be clustered in a new cluster (no overlap with initial clusters)"""
-    clusterOfE7 = self.getUpdatedClusterData()['elementToCluster']['e7']
-    elementsOfCluster = self.getUpdatedClusterData()['clusterToElement'][clusterOfE7]
-    self.assertEqual(len(elementsOfCluster), 1, msg=f'Other elements in the cluster of e7: {elementsOfCluster}')
-
-
-# -----------------------------------------------------------------------------
-class TestClusteringSingleInput(TestClustering, unittest.TestCase):
+class TestClusteringSingleInput(InitialClusteringSize, InitialElementsTogether, unittest.TestCase):
   """A concrete integration test class that executes tests of the TestClustering class for a clustering with a single input files."""
 
   # ---------------------------------------------------------------------------
@@ -104,16 +45,10 @@ class TestClusteringSingleInput(TestClustering, unittest.TestCase):
     return TestClusteringSingleInput.initialClusterData
 
   # ---------------------------------------------------------------------------
-  def getUpdatedClusterData(self):
-    return TestClusteringSingleInput.updatedClusterData
-
-  # ---------------------------------------------------------------------------
   @classmethod
   def setUpClass(cls):
     cls.tempInitialClusters = os.path.join(tempfile.gettempdir(), 'initial-clusters.csv')
-    cls.tempNewClusters = os.path.join(tempfile.gettempdir(), 'updated-clusters.csv')
 
-    print('Initial clustering ...')
     # Cluster from scratch
     #
     initialClustering(
@@ -124,9 +59,31 @@ class TestClusteringSingleInput(TestClustering, unittest.TestCase):
       delimiter=","
     )
 
-    print()
-    print('Update clusters ...')
-    # Cluster more
+    # read the script output into an internal data structure
+    #
+    cls.initialClusterData = readOutput(cls.tempInitialClusters)
+
+   # ---------------------------------------------------------------------------
+  @classmethod
+  def tearDownClass(cls):
+    if os.path.isfile(cls.tempInitialClusters):
+      os.remove(cls.tempInitialClusters)
+
+
+# -----------------------------------------------------------------------------
+class TestUpdateClusteringSingleInput(UpdateClusteringSize,UpdateClusteringElementsTogether, unittest.TestCase):
+  """A concrete integration test class that executes tests for the update of clusters with a single input files."""
+
+  # ---------------------------------------------------------------------------
+  def getUpdatedClusterData(self):
+    return TestUpdateClusteringSingleInput.updatedClusterData
+
+  # ---------------------------------------------------------------------------
+  @classmethod
+  def setUpClass(cls):
+    cls.tempNewClusters = os.path.join(tempfile.gettempdir(), 'updated-clusters.csv')
+
+    # Cluster
     #
     updateClusters(
       inputFilenames=["test/resources/cluster-input-2.csv"],
@@ -140,30 +97,56 @@ class TestClusteringSingleInput(TestClustering, unittest.TestCase):
 
     # read the script output into an internal data structure
     #
-    cls.initialClusterData = readOutput(cls.tempInitialClusters)
     cls.updatedClusterData = readOutput(cls.tempNewClusters)
 
    # ---------------------------------------------------------------------------
   @classmethod
   def tearDownClass(cls):
-    if os.path.isfile(cls.tempInitialClusters):
-      os.remove(cls.tempInitialClusters)
     if os.path.isfile(cls.tempNewClusters):
       os.remove(cls.tempNewClusters)
 
 
 
 # -----------------------------------------------------------------------------
-class TestClusteringMultipleInput(TestClustering, unittest.TestCase):
-  """A concrete integration test class that executes tests of the TestClustering class for a clustering with several input files."""
+class TestClusteringMultipleInput(InitialClusteringSize, InitialElementsTogether, unittest.TestCase):
+  """A concrete integration test class that executes tests for a clustering with several input files."""
 
   # ---------------------------------------------------------------------------
   def getInitialClusterData(self):
     return TestClusteringMultipleInput.initialClusterData
 
   # ---------------------------------------------------------------------------
+  @classmethod
+  def setUpClass(cls):
+    cls.tempInitialClusters = os.path.join(tempfile.gettempdir(), 'initial-clusters-multiple-input-files.csv')
+
+    # Cluster from scratch
+    #
+    initialClustering(
+      inputFilenames=["test/resources/cluster-input-1.1.csv", "test/resources/cluster-input-1.2.csv"],
+      outputFilename=cls.tempInitialClusters,
+      idColumnName="elementID",
+      keyColumnName="descriptiveKey",
+      delimiter=","
+    )
+
+    # read the script output into an internal data structure
+    #
+    cls.initialClusterData = readOutput(cls.tempInitialClusters)
+
+   # ---------------------------------------------------------------------------
+  @classmethod
+  def tearDownClass(cls):
+    if os.path.isfile(cls.tempInitialClusters):
+      os.remove(cls.tempInitialClusters)
+
+# -----------------------------------------------------------------------------
+class TestClusteringOverlappingKeysDifferentClustersReusingKeys(UpdateClusteringSize, UpdateClusteringElementsTogether, unittest.TestCase):
+  """A concrete integration test class that executes tests of the TestClustering class for a clustering with a single input file an."""
+
+  # ---------------------------------------------------------------------------
   def getUpdatedClusterData(self):
-    return TestClusteringMultipleInput.updatedClusterData
+    return TestClusteringOverlappingKeysDifferentClustersReusingKeys.updatedClusterData
 
   # ---------------------------------------------------------------------------
   @classmethod
@@ -208,6 +191,7 @@ class TestClusteringMultipleInput(TestClustering, unittest.TestCase):
       os.remove(cls.tempInitialClusters)
     if os.path.isfile(cls.tempNewClusters):
       os.remove(cls.tempNewClusters)
+
